refactor(server): extract omp adapter collaborators into a feature module with barrel
