+++ apps/server/src/orchestration/Layers/CheckpointReactor.test.ts
@@ -55,7 +55,7 @@ import {
   ProviderService,
   type ProviderServiceShape,
 } from "../../provider/Services/ProviderService.ts";
-import { OmpAdapter } from "../../provider/omp/OmpAdapter.ts";
+import { OmpAdapter } from "../../provider/omp/index.ts";
 import { checkpointRefForThreadTurn } from "../../checkpointing/Utils.ts";
 import { ServerConfig } from "../../config.ts";
 import * as WorkspaceEntries from "../../workspace/WorkspaceEntries.ts";
