+++ apps/server/src/provider/Layers/ProviderService.ts
@@ -49,7 +49,7 @@ import {
   withMetrics,
 } from "../../observability/Metrics.ts";
 import { type ProviderAdapterError, ProviderValidationError } from "../Errors.ts";
-import { OmpAdapter } from "../omp/OmpAdapter.ts";
+import { OmpAdapter } from "../omp/index.ts";
 import type { ProviderAdapterShape } from "../Services/ProviderAdapter.ts";
 import * as ProviderAdapterRegistry from "../Services/ProviderAdapterRegistry.ts";
 import * as ProviderService from "../Services/ProviderService.ts";
