+++ apps/server/src/provider/omp/OmpAdapter.ts
@@ -22,9 +22,10 @@
  *
  * @module provider/omp/OmpAdapter
  */
+import { ReviewBlockDecoder } from "./ReviewBlockDecoder.ts";
+import { OmpToolPresentation } from "./OmpToolPresentation.ts";
 import {
   type ApprovalRequestId,
-  type CanonicalItemType,
   EventId,
   type ProviderApprovalDecision,
   type ProviderRuntimeEvent,
@@ -41,19 +42,14 @@ import {
   RuntimeTaskId,
   type RuntimeMode,
   type ProviderTurnInteractionMode,
-  PullRequestDiffSide,
-  PositiveInt,
   ReviewFinding,
-  ReviewFindingSeverity,
   ReviewRunVerdict,
-  TrimmedNonEmptyString,
   type ThreadId,
   TurnId,
 } from "@t3tools/contracts";
 import * as Cause from "effect/Cause";
 import * as Clock from "effect/Clock";
 import * as DateTime from "effect/DateTime";
-import * as Option from "effect/Option";
 import * as Effect from "effect/Effect";
 import * as Exit from "effect/Exit";
 import * as Queue from "effect/Queue";
@@ -85,40 +81,12 @@ const isProviderAdapterProcessError = Schema.is(ProviderAdapterProcessError);
 const isOmpSpawnError = Schema.is(OmpSpawnError);
 
 // ---------------------------------------------------------------------------
-// Review mode (issue #42): a review turn ends with a fenced JSON findings
-// block. The adapter decodes it and emits one `review.finding` runtime event
-// per finding; a missing or malformed block fails the turn so the review run
-// surfaces the error instead of silently dropping findings.
+// Review findings block decoding (issue #42): the persona ends with one
+// fenced JSON block. The adapter decodes it and emits one `review.finding`
+// runtime event per finding; a missing or malformed block fails the turn so
+// the review run surfaces the error instead of silently dropping findings.
+// Decoding lives in ReviewBlockDecoder (this file keeps the emit orchestration).
 // ---------------------------------------------------------------------------
-
-const ReviewFindingInputSchema = Schema.Struct({
-  file: TrimmedNonEmptyString,
-  message: TrimmedNonEmptyString,
-  line: Schema.NullOr(PositiveInt).pipe(Schema.withDecodingDefault(Effect.succeed(null))),
-  side: PullRequestDiffSide.pipe(Schema.withDecodingDefault(Effect.succeed("right"))),
-  severity: ReviewFindingSeverity.pipe(Schema.withDecodingDefault(Effect.succeed("should-fix"))),
-  symbol: Schema.NullOr(TrimmedNonEmptyString).pipe(
-    Schema.withDecodingDefault(Effect.succeed(null)),
-  ),
-});
-type ReviewFindingInput = typeof ReviewFindingInputSchema.Type;
-
-const ReviewFindingsBlockSchema = Schema.Struct({
-  findings: Schema.Array(ReviewFindingInputSchema),
-  verdict: Schema.optional(ReviewRunVerdict),
-  summary: Schema.optional(TrimmedNonEmptyString),
-  filesReviewed: Schema.optional(Schema.Array(TrimmedNonEmptyString)),
-});
-
-const decodeReviewFindingsBlock = Schema.decodeUnknownOption(ReviewFindingsBlockSchema);
-
-function parseJsonBlock(text: string): unknown {
-  try {
-    return JSON.parse(text) as unknown;
-  } catch {
-    return undefined;
-  }
-}
 
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null && !Array.isArray(value);
@@ -247,6 +215,8 @@ export class OmpAdapter {
   readonly #randomUUID: Effect.Effect<string>;
   readonly #resolveRoleModel: OmpResolveRoleModel;
   readonly #capabilitiesService: OmpAdapterOptions["capabilitiesService"];
+  readonly #reviewBlockDecoder = new ReviewBlockDecoder();
+  readonly #toolPresentation = new OmpToolPresentation();
 
   public constructor(
     runtime: OmpRpcClient,
@@ -1258,7 +1228,7 @@ export class OmpAdapter {
     if (typeof frame.toolCallId !== "string") {
       return Effect.void;
     }
-    const delta = formatOmpToolOutputText(frame.partialResult);
+    const delta = this.#toolPresentation.formatOutputText(frame.partialResult);
     if (delta.length === 0) {
       return Effect.void;
     }
@@ -1287,7 +1257,7 @@ export class OmpAdapter {
     const tracked = session.toolCalls.get(toolCallId);
     const args = frame.args !== undefined ? frame.args : tracked?.args;
     const intent = typeof frame.intent === "string" ? frame.intent : tracked?.intent;
-    const presentation = presentOmpToolCall({
+    const presentation = this.#toolPresentation.present({
       toolCallId,
       toolName,
       args,
@@ -1328,7 +1298,7 @@ export class OmpAdapter {
       args: input.args,
       intent: input.intent,
     });
-    const presentation = presentOmpToolCall(input);
+    const presentation = this.#toolPresentation.present(input);
     return this.#emit({
       type: "item.started",
       threadId: session.threadId,
@@ -1460,42 +1430,19 @@ export class OmpAdapter {
     readonly filesReviewed?: ReadonlyArray<string>;
   } | null> {
     return Effect.gen({ self: this }, function* () {
-      if (runText === null || runText.length === 0) {
-        return null;
-      }
-      // The persona ends with one fenced JSON block; take the last one so any
-      // incidental prose fences earlier in the review are ignored.
-      const fence = /```json\s*([\s\S]*?)```/gu;
-      let last: RegExpExecArray | null = null;
-      let match: RegExpExecArray | null;
-      while ((match = fence.exec(runText)) !== null) {
-        last = match;
-      }
-      if (last === null) {
-        return null;
-      }
-      const blockText = last[1];
-      if (blockText === undefined) {
-        return null;
-      }
-      const parsed = parseJsonBlock(blockText);
-      if (parsed === undefined) {
-        return null;
-      }
-      const decoded = decodeReviewFindingsBlock(parsed);
-      if (Option.isNone(decoded)) {
+      const decoded = this.#reviewBlockDecoder.decode(runText);
+      if (decoded === null) {
         return null;
       }
       const findings: ReviewFinding[] = [];
-      for (const entry of decoded.value.findings as ReadonlyArray<ReviewFindingInput>) {
+      for (const entry of decoded.findings) {
         findings.push({ id: `finding-${yield* this.#randomUUID}`, ...entry });
       }
-      const { verdict, summary, filesReviewed } = decoded.value;
       return {
         findings,
-        ...(verdict === undefined ? {} : { verdict }),
-        ...(summary === undefined ? {} : { summary }),
-        ...(filesReviewed === undefined ? {} : { filesReviewed }),
+        ...(decoded.verdict === undefined ? {} : { verdict: decoded.verdict }),
+        ...(decoded.summary === undefined ? {} : { summary: decoded.summary }),
+        ...(decoded.filesReviewed === undefined ? {} : { filesReviewed: decoded.filesReviewed }),
       };
     });
   }
@@ -2079,152 +2026,6 @@ function runtimeTaskStatusFromOmpProgress(status: unknown): RuntimeTaskStatus | 
 
 function isLocalOnlyPromptResponse(response: object): boolean {
   return isRecord(response) && isRecord(response.data) && response.data.agentInvoked === false;
-}
-
-function ompToolKind(toolName: string): string | undefined {
-  const name = toolName.toLowerCase();
-  if (name === "bash" || name === "shell" || name === "execute") {
-    return "execute";
-  }
-  if (name === "read") {
-    return "read";
-  }
-  if (
-    name === "write" ||
-    name === "edit" ||
-    name === "multiedit" ||
-    name.includes("patch") ||
-    name.includes("edit")
-  ) {
-    return "edit";
-  }
-  if (name.includes("web_search") || name === "websearch" || name === "grep" || name === "find") {
-    return "search";
-  }
-  return undefined;
-}
-
-function ompToolItemType(toolName: string): CanonicalItemType {
-  const name = toolName.toLowerCase();
-  const kind = ompToolKind(toolName);
-  if (kind === "execute") {
-    return "command_execution";
-  }
-  if (kind === "edit") {
-    return "file_change";
-  }
-  if (name.includes("web_search") || name === "websearch") {
-    return "web_search";
-  }
-  if (name.startsWith("mcp__") || name.startsWith("mcp_")) {
-    return "mcp_tool_call";
-  }
-  if (name === "task" || name.includes("collab") || name === "agent") {
-    return "collab_agent_tool_call";
-  }
-  // read + other omp tools: dynamic_tool_call + data.kind for UI presentation
-  return "dynamic_tool_call";
-}
-
-function extractOmpToolCommand(args: unknown): string | undefined {
-  if (!isRecord(args) || typeof args.command !== "string") {
-    return undefined;
-  }
-  const command = args.command.trim();
-  return command.length > 0 ? command : undefined;
-}
-
-function extractOmpToolPath(args: unknown): string | undefined {
-  if (!isRecord(args)) {
-    return undefined;
-  }
-  for (const key of ["path", "filePath", "filename", "file", "target"] as const) {
-    const value = args[key];
-    if (typeof value === "string" && value.trim().length > 0) {
-      return value.trim();
-    }
-  }
-  return undefined;
-}
-
-/** Unwrap omp AgentToolResult `{ content: [{ type:"text", text }] }` to plain text. */
-function formatOmpToolOutputText(value: unknown): string {
-  if (value === undefined || value === null) {
-    return "";
-  }
-  if (typeof value === "string") {
-    return value;
-  }
-  if (typeof value === "number" || typeof value === "boolean") {
-    return String(value);
-  }
-  if (isRecord(value) && Array.isArray(value.content)) {
-    const parts: string[] = [];
-    for (const entry of value.content) {
-      if (isRecord(entry) && typeof entry.text === "string" && entry.text.length > 0) {
-        parts.push(entry.text);
-      }
-    }
-    if (parts.length > 0) {
-      return parts.join("");
-    }
-  }
-  if (isRecord(value)) {
-    if (typeof value.stdout === "string" && value.stdout.length > 0) {
-      return value.stdout;
-    }
-    if (typeof value.text === "string" && value.text.length > 0) {
-      return value.text;
-    }
-  }
-  return "";
-}
-
-function truncateDetail(value: string, maxLength = 500): string {
-  if (value.length <= maxLength) {
-    return value;
-  }
-  return `${value.slice(0, maxLength - 1)}…`;
-}
-
-function presentOmpToolCall(input: {
-  readonly toolCallId: string;
-  readonly toolName: string;
-  readonly args: unknown;
-  readonly intent?: string | undefined;
-  readonly result?: unknown;
-  readonly isError?: boolean;
-}): {
-  readonly itemType: CanonicalItemType;
-  readonly title: string;
-  readonly detail: string | undefined;
-  readonly data: Record<string, unknown>;
-} {
-  const kind = ompToolKind(input.toolName);
-  const command = extractOmpToolCommand(input.args);
-  const path = extractOmpToolPath(input.args);
-  const intent =
-    typeof input.intent === "string" && input.intent.trim().length > 0
-      ? input.intent.trim()
-      : undefined;
-  const detailSource = intent ?? command ?? path;
-  const outputText = formatOmpToolOutputText(input.result);
-  return {
-    itemType: ompToolItemType(input.toolName),
-    title: input.toolName,
-    detail: detailSource !== undefined ? truncateDetail(detailSource) : undefined,
-    data: {
-      toolCallId: input.toolCallId,
-      toolName: input.toolName,
-      ...(kind === undefined ? {} : { kind }),
-      ...(command === undefined ? {} : { command }),
-      ...(input.args === undefined ? {} : { rawInput: input.args, args: input.args }),
-      ...(path === undefined ? {} : { locations: [{ path }] }),
-      ...(outputText.length > 0 ? { rawOutput: { content: outputText } } : {}),
-      ...(input.result === undefined ? {} : { result: input.result }),
-      ...(input.isError === undefined ? {} : { isError: input.isError }),
-    },
-  };
 }
 
 function loginProvidersFromResponse(
